fix(tests): moved db connection from setup to general beforeAll block
