+++ shared/jest.integration-tests-setup.ts
@@ -1,11 +1,7 @@
-import * as path from 'path';
-
 import { config as dotenvConfig } from 'dotenv';
 import { DockerComposeEnvironment, Wait } from 'testcontainers';
 import { registerTsProject } from '@nx/js/src/internal';
 const cleanupRegisteredPaths = registerTsProject('./tsconfig.base.json');
-
-import { getDb } from '@nrcno/core-db';
 
 export default async function () {
   // Start services that the app needs to run (e.g. database, docker-compose, etc.).
@@ -24,23 +20,6 @@ export default async function () {
     .withWaitStrategy('db', Wait.forHealthCheck())
     .withStartupTimeout(120 * 1000)
     .up();
-
-  // Initialise the database connection
-  const config = {
-    host: process.env.DB_HOST,
-    user: process.env.DB_USER,
-    password: process.env.DB_PASSWORD,
-    database: process.env.DB_NAME,
-  };
-  const db = getDb(config);
-  await db.migrate.latest({
-    loadExtensions: ['.js'],
-  });
-  await db.seed.run({
-    loadExtensions: ['.js'],
-    directory: path.join(db.client.config.seeds.directory, 'common'),
-  });
-  (global as any).db = db;
 
   // Pass the environment to global teardown
   (global as any).__ENVIRONMENT__ = environment;
